+++ changelog.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.2] - 2024-06-12
+
+- Corrigidos campos url de Episode e Location para unique
+- Correção na views de Episode e Location de acordo com a mudança
+
 ## [1.0.1] - 2024-06-12
 
 - Corrigidos campos de Episode e Location para many to many